requirements.txt
livekit-agents>=1.0
livekit-plugins-deepgram
livekit-plugins-openai
redis>=5.0
